--- src/components/missions/Missions.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const Missions = () => <div>Missions</div>;
-
-export default Missions;
