+++ src/components/spaceHub/SpaceHub.js
@@ -1,6 +1,6 @@
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import Navigation from '../Navigation/Navigation';
-import Missions from '../missions/Missions';
+import Missions from '../missions/MissionsContainer';
 import Rockets from '../rocket/Rockets';
 import 'bootstrap/dist/css/bootstrap.min.css';
